+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 # UNRELEASED
 
+## Changed
+
+* `Term` prettyprinter now uses `Name`s instead of `DeBruijn`s for readability
+
 # 1.13.0 -- 04-03-2026
 
 ## Added